fixed some stuff
ok, so nate's code wasn't working for me, so here's a bunch of stuff i
did to make it work for @latourliturgies

* merged in some changes that jacob harris made since nate forked his
project (htmlentities, blank tweet protection)

* there was some stuff that ebook.rb needs that wasn't specified in
twitter_init.rb so i added it

* commented some stuff in ebook.rb that had been confusing to me

* added some debugging code to the tweet generation in ebook.rb so it's
clear what is being scraped from which account

i'm pretty sure these changes are helpful, but i'm not sure if i'm
contributing them in a helpful way, since this is my first time trying
to contribute to someone else's code on github. apologies in advance!